Removed style in library
Style in library is not necessary. It has side effect in case that
application is composed from library that contains all code, styles,
graphics, ... and application that uses this library and defines manifest
(way to create lite/full versions of same application). In this case
HorizontalVariableListView can overwrite style defined by the application
library.